# openapi.yaml
openapi: 3.1.0
info:
  title: Onikişubat Belediyesi Atık Yönetimi API
  version: 1.0.0
  description: Cookie tabanlı oturum, CSRF koruması ve rol bazlı yetkilendirme kullanan API sözleşmesi.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /konteynerler:
    get:
      security: []
      summary: Aktif konteynerleri sayfalı listeler
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses: { '200': { $ref: '#/components/responses/Paginated' } }
  /sikayetler:
    post:
      security: []
      summary: Vatandaş şikâyeti ve en fazla üç fotoğraf oluşturur
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [vatandas_ad_soyad, vatandas_telefon, konteyner_id, sikayet_turu, sikayet_metni]
              properties:
                vatandas_ad_soyad: { type: string, minLength: 3 }
                vatandas_telefon: { type: string, pattern: '^05[0-9]{9}$' }
                konteyner_id: { type: integer, minimum: 1 }
                sikayet_turu: { type: string, enum: [kati_atik, geri_donusum] }
                sikayet_kategorisi: { type: string }
                sikayet_metni: { type: string }
                fotograflar: { type: array, maxItems: 3, items: { type: string, format: binary } }
      responses: { '201': { description: Şikâyet oluşturuldu }, '400': { $ref: '#/components/responses/BadRequest' } }
    get:
      summary: Yönetici için şikâyet listesi
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Paginated' }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /sofor/konteynerler:
    get:
      summary: Şoförün çevrimdışı önbelleğe alınabilen görevleri
      x-required-role: sofor
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/Paginated' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /sofor/toplama-kayitlari:
    post:
      summary: İdempotent toplama veya atlama kaydı
      x-required-role: sofor
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string, minLength: 20, maxLength: 64, pattern: '^[A-Za-z0-9-]+$' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CollectionRequest' } } }
      responses: { '201': { description: Kayıt oluşturuldu veya aynı anahtarın mevcut kaydı döndürüldü }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /sofor/gorevler:
    get:
      summary: Oturumdaki şoförün görevlerini öncelik sırasıyla listeler
      x-required-role: sofor
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { name: durum, in: query, schema: { type: string, enum: [atandi, devam_ediyor, tamamlandi, atlandi, iptal_edildi] } }]
      responses: { '200': { $ref: '#/components/responses/Paginated' } }
  /sofor/gorevler/{id}/baslat:
    patch:
      summary: Oturumdaki şoförün atanmış görevini başlatır
      x-required-role: sofor
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: Görev başlatıldı }, '404': { description: Başlatılabilir görev bulunamadı } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: auth_token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
  schemas:
...
    CollectionRequest:
      type: object
      required: [konteyner_id, durum]
      properties: { konteyner_id: { type: integer, minimum: 1 }, durum: { type: string, enum: [toplandi, atlanildi] }, sebep: { type: string }, diger_aciklama: { type: string } }
...
  responses:
    Paginated: { description: Sayfalı başarılı yanıt }
    BadRequest: { description: Geçersiz istek }
...
    Forbidden: { description: Rol bu işlem için yetkili değil }
...